Player class added
Added a player class and replaced movement with methods for player class, player class is not finished. Moved the assets folder because that fixes the old issue me and luke had loading the files, which only seemed to work with lewis originially so deal with it. Now has "assets/image" as I decided to cleanly keep the assets in a folder still.